Build with tshy
