+++ website/.eslintrc.json
@@ -6,5 +6,17 @@
   },
   "rules": {
     "n/no-missing-require": 0
-  }
+  },
+  "overrides": [
+    {
+      "files": ["*.config.js"],
+      "globals": {
+        "require": "readonly",
+        "module": "readonly"
+      },
+      "parserOptions": {
+        "sourceType": "script"
+      }
+    }
+  ]
 }
