+++ website/docs/advanced/configuring-cheerio.md
@@ -144,7 +144,7 @@ You can also use Cheerio's _slim_ export, which always uses `htmlparser2`. This
 avoids loading `parse5`, which saves some bytes eg. in browser environments:
 
 ```js
-import * as cheerio from 'cheerio/lib/slim';
+import * as cheerio from 'cheerio/slim';
 ```
 
 :::
